+++ services/src/main/java/org/keycloak/broker/oidc/OIDCIdentityProviderConfig.java
@@ -34,6 +34,7 @@ public class OIDCIdentityProviderConfig extends OAuth2IdentityProviderConfig {
     public static final String IS_ACCESS_TOKEN_JWT = "isAccessTokenJWT";
     public static final String TOKEN_INTROSPECTION_URL = "tokenIntrospectionUrl";
     public static final String VALIDATE_REFRESH_TOKEN = "validateRefreshToken";
+    public static final String CLAIMS_PARAMETER_SUPPORTED = "claimsParameterSupported";
 
     public OIDCIdentityProviderConfig(IdentityProviderModel identityProviderModel) {
         super(identityProviderModel);
@@ -201,6 +202,21 @@ public class OIDCIdentityProviderConfig extends OAuth2IdentityProviderConfig {
         getConfig().put(VALIDATE_REFRESH_TOKEN, String.valueOf(validateRefreshToken));
     }
 
+    public void setPassSetMfa(boolean passSetMfa) {
+        getConfig().put(IdentityProviderModel.PASS_SET_MFA, String.valueOf(passSetMfa));
+    }
+
+    public boolean isPassSetMfa() {
+        return Boolean.valueOf(getConfig().get(IdentityProviderModel.PASS_SET_MFA));
+    }
+
+    public void setClaimsParameterSupported(boolean claimsParameterSupported) {
+        getConfig().put(CLAIMS_PARAMETER_SUPPORTED, String.valueOf(claimsParameterSupported));
+    }
+
+    public boolean isClaimsParameterSupported() {
+        return Boolean.valueOf(getConfig().get(CLAIMS_PARAMETER_SUPPORTED));
+    }
 
     @Override
     public void validate(RealmModel realm) {