+++ testsuite/integration-arquillian/tests/base/src/test/java/org/keycloak/testsuite/admin/IdentityProviderTest.java
@@ -1274,7 +1274,7 @@ public class IdentityProviderTest extends AbstractAdminTest {
             map.put("fromUrl", "http://localhost:8880/oidc-idp");
 
             Map<String, String> result = realm.identityProviders().importFrom(map);
-            assertThat(result.keySet(), containsInAnyOrder("issuer", "authorizationUrl", "tokenUrl", "userInfoUrl", "validateSignature", "useJwksUrl", "jwksUrl"));
+            assertThat(result.keySet(), containsInAnyOrder("issuer", "authorizationUrl", "tokenUrl", "userInfoUrl", "validateSignature", "useJwksUrl", "jwksUrl","claimsParameterSupported"));
             assertThat(result, hasEntry("authorizationUrl", "https://aai.egi.eu/oidc/authorize"));
             assertThat(result, hasEntry("tokenUrl", "https://aai.egi.eu/oidc/token"));
 
@@ -1328,7 +1328,7 @@ public class IdentityProviderTest extends AbstractAdminTest {
     }
 
     private void assertOidcConfig(Map<String, String> config, boolean hasExecuted) {
-        Set fields = Stream.of("issuer", "authorizationUrl", "tokenUrl", "userInfoUrl", "validateSignature", "useJwksUrl", "jwksUrl", "autoUpdate", "metadataUrl", "refreshPeriod").collect(Collectors.toSet());
+        Set fields = Stream.of("issuer", "authorizationUrl", "tokenUrl", "userInfoUrl", "validateSignature", "useJwksUrl", "jwksUrl", "autoUpdate", "metadataUrl", "refreshPeriod","claimsParameterSupported").collect(Collectors.toSet());
         //autoupdated has been executed -  add lastRefreshTime
         if (hasExecuted)
             fields.add("lastRefreshTime");