+++ CHANGELOG.md
@@ -10,6 +10,11 @@ Full Keycloak upstream jira issue can be shown if filtered by Fix version.
 
 Our Keycloak version is working well with PostgreSQL database. For using other SQL databases, text field in database need to be evaluated.
 
+## [Unreleased]
+
+### Added
+- Signalling Multi-Factor Authentication (MFA) requests to OIDC and SAML Identity Providers 
+
 ## [22.0.13-1.14] - 2025-02-13
 
 ### Changed